services/ingestor/requirements.txt
requests==2.32.3
kafka-python==2.0.2
python-dotenv==1.0.1